+++ DungeonKit/5e/Classes/DKClass5E.m
@@ -19,10 +19,10 @@
 
 #pragma mark -
 
-+ (DKModifierGroup*)abilityScoreImprovementForThreshold:(int)threshold level:(DKNumericStatistic*)classLevel {
++ (DKModifierGroup*)abilityScoreImprovementForThreshold:(NSInteger)threshold level:(DKNumericStatistic*)classLevel {
     
     DKModifierGroup* abilityScoreSubgroup = [[DKModifierGroup alloc] init];
-    abilityScoreSubgroup.explanation = [NSString stringWithFormat:@"Ability score improvements for level %i", threshold];
+    abilityScoreSubgroup.explanation = [NSString stringWithFormat:@"Ability score improvements for level %li", (long)threshold];
     
     DKDependentModifier* strModifier = [[DKDependentModifier alloc] initWithSource:classLevel
                                                                           value:[NSExpression expressionForConstantValue:@(1)]
@@ -30,7 +30,7 @@
                                                                                                     isGreaterThanOrEqualTo:threshold]
                                                                        priority:kDKModifierPriority_Additive
                                                                      expression:[DKModifierBuilder simpleAdditionModifierExpression]];
-    strModifier.explanation = [NSString stringWithFormat:@"Ability score improvement for level %i (default)", threshold];
+    strModifier.explanation = [NSString stringWithFormat:@"Ability score improvement for level %li (default)", (long)threshold];
     [abilityScoreSubgroup addModifier:strModifier forStatisticID:DKStatIDStrength];
     
     DKDependentModifier* dexModifier = [[DKDependentModifier alloc] initWithSource:classLevel
@@ -39,13 +39,13 @@
                                                                                                     isGreaterThanOrEqualTo:threshold]
                                                                           priority:kDKModifierPriority_Additive
                                                                         expression:[DKModifierBuilder simpleAdditionModifierExpression]];
-    dexModifier.explanation = [NSString stringWithFormat:@"Ability score improvement for level %i (default)", threshold];
+    dexModifier.explanation = [NSString stringWithFormat:@"Ability score improvement for level %li (default)", (long)threshold];
     [abilityScoreSubgroup addModifier:dexModifier forStatisticID:DKStatIDDexterity];
     
     return abilityScoreSubgroup;
 }
 
-+ (DKModifier*)hitDiceModifierForSides:(int)sides level:(DKNumericStatistic*)classLevel {
++ (DKModifier*)hitDiceModifierForSides:(NSInteger)sides level:(DKNumericStatistic*)classLevel {
     
     NSExpression* value = [NSExpression expressionForFunction:[NSExpression expressionForConstantValue: [DKDiceCollection diceCollection]]
                                                  selectorName:@"diceByAddingQuantity:sides:"
